MASSIVE refactoring - please do not update your blogs yet
This is the refactoring I've been working on. Honk is now a modular Sinatra app
that runs with rackup. Some parts are still missing (notably the comment
posting), and there may be bugs. I hope this version will be more usable, and
that it will play nice with other Rack apps.

*Please, do not use versions up from this one until further notice! They are
prone to bugs and many things are missing.*